backend/requirements.txt
### backend/requirements.txt
fastapi
uvicorn
pymongo
boto3
python-dotenv
PyMuPDF
sentence-splitter
pytest